Step 2/6: Implement strongly-typed C# models: update `Models/DashboardData.cs`, `M...
